Add support to reuse existing volume
1. Adds two new CLI subcommands:
- --reuse-storage-units
- --volume-id
2. Handles "File exists" error by giving option to reuse by
- overwriting volume-id
- or by passing existing volume-id

Fixes: #148, #149

+++ mgr/src/server/plugins/volume_utils.cr
@@ -66,20 +66,63 @@ def validate_volume_create(req)
         return NodeResponse.new(false, {"error": "Storage unit parent directory(#{Path[storage_unit.path].parent}) not exists"}.to_json)
       end
 
+      storage_unit_pre_exists = File.exists?(Path[storage_unit.path])
+
+      puts "storage_units_pre_exists: #{storage_unit_pre_exists}"
+      puts "reuse storage_units: #{req.reuse_storage_units}"
+      puts "volume id: #{req.id}"
+
+
+      # Validation to check if creation of storage-unit & xattr are supported or not.
       begin
         Dir.mkdir storage_unit.path
-      rescue ex : Exception
-        return NodeResponse.new(false, {"error": "Failed to create Storage unit path #{storage_unit.path} (Error: #{ex})"}.to_json)
-      end
-
-      begin
         xattr = XAttr.new(storage_unit.path)
         xattr[TEST_XATTR_NAME] = TEST_XATTR_VALUE
+      rescue ex : Exception
+        return NodeResponse.new(false, {"error": "Failed to create Storage unit path #{storage_unit.path} (Error: #{ex})"}.to_json)
       rescue ex : IO::Error
         return NodeResponse.new(false, {"error": "Extended attributes are not supported for #{storage_unit.path} (Error: #{ex})"}.to_json)
       ensure
-        FileUtils.rmdir storage_unit.path
+        FileUtils.rmdir storage_unit.path unless storage_unit_pre_exists
       end
+
+      begin
+        Dir.mkdir storage_unit.path
+      rescue ex: Exception
+        xattr = XAttr.new(storage_unit.path)
+        if xattr[VOLUME_ID_XATTR_NAME]
+
+          # if xattr[VOLUME_ID_XATTR_NAME] == req.id
+          #   puts "pass: matching volume id"
+
+          if xattr[VOLUME_ID_XATTR_NAME] != req.id && req.reuse_storage_units == true
+            #Overwrite existing volume-id xattr
+            xattr[VOLUME_ID_XATTR_NAME] = req.id
+
+          else if xattr[VOLUME_ID_XATTR_NAME] != req.id
+            return NodeResponse.new(false, {"error": "Extended attributes & volume-id does not match for #{storage_unit.path}"}.to_json)
+
+          else
+            return NodeResponse.new(false, {"error": "Failed to create storage-unit path #{storage_unit.path}. It is already part of another volume"}.to_json)
+          end
+
+        end
+      end
+    end
+
+
+        #else
+        # puts "pass: no xattr"
+
+
+      # begin
+      #   xattr = XAttr.new(storage_unit.path)
+      #   xattr[TEST_XATTR_NAME] = TEST_XATTR_VALUE
+      # rescue ex : IO::Error
+      #   return NodeResponse.new(false, {"error": "Extended attributes are not supported for #{storage_unit.path} (Error: #{ex})"}.to_json)
+      # ensure
+      #   FileUtils.rmdir storage_unit.path
+      # end
     end
   end
 
@@ -122,19 +165,18 @@ def handle_volume_create(data, stopped = false)
       next unless storage_unit.node.name == GlobalConfig.local_node.name
 
       # Create the Storage Unit
-      Dir.mkdir storage_unit.path
+      Dir.mkdir_p storage_unit.path
 
       # Set volume-id xattr, ignore if same Volume ID exists
       volume_id = UUID.new(req.id)
       begin
         xattr = XAttr.new(storage_unit.path, only_create: true)
         xattr[VOLUME_ID_XATTR_NAME] = volume_id.bytes.to_slice
+        puts xattr[VOLUME_ID_XATTR_NAME]
       rescue ex : IO::Error
-        if ex.os_error == Errno::EEXIST && xattr.not_nil![VOLUME_ID_XATTR_NAME] != volume_id.bytes.to_slice
-          return NodeResponse.new(false, {"error": "Storage Unit #{storage_unit.node.name}:#{storage_unit.path} is already used with another Volume"}.to_json)
-        else
+        if ex.os_error != Errno::EEXIST
           return NodeResponse.new(false, {"error": "Failed to set Volume ID Xattr. Error=#{ex}"}.to_json)
-        end
+       end
       end
 
       # Create Meta directories
